% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{cat2}
\alias{cat2}
\title{Color Stdout}
\usage{
cat2(..., level = "DEBUG", print_level = FALSE, file = "",
  sep = " ", fill = FALSE, labels = NULL, append = FALSE,
  pal = list(DEBUG = "grey60", INFO = "#1d9f34", WARNING = "#ec942c",
  ERROR = "#f02c2c", FATAL = "#763053", DEFAULT = "#000000"))
}
\arguments{
\item{...}{to be printed}

\item{level}{DEBUG, INFO, WARNING, ERROR, or FATAL (total 5 levels)}

\item{print_level}{if true, prepend levels before messages}

\item{file}{pass to base::cat}

\item{sep}{pass to base::cat}

\item{fill}{pass to base::cat}

\item{labels}{pass to base::cat}

\item{append}{pass to base::cat}

\item{pal}{a named list defining colors see details}
}
\description{
Color Stdout
}
\details{
There are five levels by default: DEBUG, INFO, WARNING, ERROR, or FATAL.
Default colors to these levels are: DEBUG (grey60), INFO (#1d9f34), WARNING
(#ec942c), ERROR (#f02c2c), FATAL (#763053) and DEFAULT (#000000, black).
If level is not in pre-set five levels, the color will be "default"-black
color.
}
