% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_packages.R
\name{check_installed_packages}
\alias{check_installed_packages}
\title{Check if packages are installed, returns missing packages}
\usage{
check_installed_packages(pkgs, libs = base::.libPaths(),
  auto_install = FALSE, ...)
}
\arguments{
\item{pkgs}{vector of packages to install}

\item{libs}{paths of libraries}

\item{auto_install}{automatically install packages if missing}

\item{...}{other params for \code{install.packages}}
}
\description{
Check if packages are installed, returns missing packages
}
