% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{create_frames}
\alias{create_frames}
\title{add decoration (frames) to plots based on their layout location}
\usage{
create_frames(
  layout,
  common = NULL,
  bottom = NULL,
  left = NULL,
  top = NULL,
  right = NULL,
  env = parent.frame(),
  exclude = NULL
)
}
\arguments{
\item{layout}{layout matrix}

\item{common}{common R expression}

\item{bottom, left, top, right}{expressions to evaluate at bottom, left,
top, and right axes}

\item{env}{environment to evaluate expressions}

\item{exclude}{sub-figure to exclude axes}
}
\description{
add decoration (frames) to plots based on their layout location
}
