% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{deparse_svec}
\alias{deparse_svec}
\title{Convert Integer Vectors To String

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}}
\usage{
deparse_svec(nums, connect = "-", concatenate = T, collapse = ",",
  max_lag = 1)
}
\description{
Convert Integer Vectors To String

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
\examples{
\dontrun{
deparse_svec(c(1:10, 15:18))
}
}
