% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{do_aggregate}
\alias{do_aggregate}
\title{Make aggregate magrittr compatible
(stable)}
\usage{
do_aggregate(data, formula, FUN, ...)
}
\arguments{
\item{data}{aggregateable data}

\item{formula}{a formula, such as y ~ x or cbind(y1, y2) ~ x1 + x2, where the y variables are numeric data to be split into groups according to the grouping x variables (usually factors).}

\item{FUN}{a function to compute the summary statistics which can be applied to all data subsets.}

\item{...}{further arguments to aggregate}
}
\description{
Make aggregate magrittr compatible
(stable)
}
\seealso{
aggregate
}
