% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ebars}
\alias{ebars}
\title{Draw Symmetric Error Bars}
\usage{
ebars(x, y = NULL, sem = NULL, length = 0.05, type = "n",
  col = "black", pt.col = col, code = 2, ...)
}
\description{
Draw Symmetric Error Bars
}
\examples{
\dontrun{
plot_clean(0:10, -1:5, xlab = 'X')
rave_axis(side = 2, at = -1:8)
ebars(x = c(2, 4), y = c(0, 3), sem = c(1, 0.5), col = c(2, 3))
}
}
