% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{F_NOOP}
\alias{F_NOOP}
\title{Apply function to input but returns input itself}
\usage{
F_NOOP(x, FUN, ...)
}
\arguments{
\item{x}{input}

\item{FUN}{function to apply on \code{x}}

\item{...}{additional parameters passing to \code{FUN}}
}
\description{
Execute the function (usually for its side
effect) then return (invisibly) the input to the function
}
