% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{getAlphaRGB}
\alias{getAlphaRGB}
\title{Get Hex Color With Transparency}
\usage{
getAlphaRGB(colname, alpha)
}
\description{
Get Hex Color With Transparency
}
\examples{
\dontrun{
getAlphaRGB('red', 0.5)
}
}
