% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{is_within}
\alias{is_within}
\alias{\%within\%}
\title{Check if a is within the range of b}
\usage{
is_within(a, b)

a \%within\% b
}
\arguments{
\item{a}{element to check (numeric)}

\item{b}{vector of numbers}
}
\description{
Check if a is within the range of b
}
