% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{lasync_expr}
\alias{lasync_expr}
\title{Apply R expressions in async settings}
\usage{
lasync_expr(.X, .expr, ..., .varname = "x", envir = parent.frame(),
  .ncore = -1)
}
\arguments{
\item{.X}{vector or list}

\item{.expr}{expression to apply}

\item{...}{passed to \code{future::future}}

\item{.varname}{variable name}

\item{envir}{environment evaluate expression}

\item{.ncore}{number of cores to use}
}
\description{
Apply R expressions in async settings
}
