% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_helpers.R
\name{get_f}
\alias{get_f}
\alias{format_f}
\alias{pretty.fres}
\title{Get statistics from linear model}
\usage{
get_f(formula, ...)

format_f(lm.mod, test_name = "All")

\method{pretty}{fres}(x, precision = 4, ...)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}
}
