% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{parse_svec}
\alias{parse_svec}
\title{Parse Text Into Numeric Vectors

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}}
\usage{
parse_svec(text, sep = ",", connect = "-:|", sort = F, unique = T)
}
\description{
Parse Text Into Numeric Vectors

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
\examples{
\dontrun{
parse_svec('1-10, 13:15,14-20')
}
}
