% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{pretty_package_docs}
\alias{pretty_package_docs}
\title{Help Formatting Documents}
\usage{
pretty_package_docs(path = ".")
}
\arguments{
\item{path}{package root path}
}
\description{
Help Formatting Documents
}
