% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{pscl}
\alias{pscl}
\title{Enforce Sum To 1, Ignoring Na In The Sum, But Keeping Them In The Output}
\usage{
pscl(x)
}
\arguments{
\item{x}{data}
}
\description{
Enforce Sum To 1, Ignoring Na In The Sum, But Keeping Them In The Output
}
