% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{rave_axis}
\alias{rave_axis}
\title{A neat way to show axis}
\usage{
rave_axis(side, at, tcl = -0.3, labels = at, las = 1,
  cex.axis = rave_cex.axis, cex.lab = rave_cex.lab, mgpy = c(3, 0.6,
  0), mgpx = c(3, 0.75, 0), ...)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
\examples{
\dontrun{
# create a blank plot with x from 0 to 10 and y from 1 to 5
plot_clean(0:10, 1:5, xlab = 'X')
rave_axis(side = 1, at = 1:8)
}
}
