% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{ruta_axis}
\alias{ruta_axis}
\title{A Neat Way To Show Axis}
\usage{
ruta_axis(
  side,
  at,
  tcl = -0.3,
  labels = at,
  las = 1,
  cex.axis = rave_cex.axis,
  cex.lab = rave_cex.lab,
  mgpy = c(3, 0.6, 0),
  mgpx = c(3, 0.75, 0),
  ...
)
}
\arguments{
\item{side}{1 to 4: bottom, left, up, right. See \code{\link{axis}}}

\item{at, tcl, labels, las, mgpy, mgpx, ...}{passed to \code{\link{axis}}}

\item{cex.axis}{tick size}

\item{cex.lab}{label size}
}
\description{
A Neat Way To Show Axis
}
\examples{
\dontrun{
# create a blank plot with x from 0 to 10 and y from 1 to 5
plot_clean(0:10, 1:5, xlab = 'X')
ruta_axis(side = 1, at = 1:8)
}
}
