% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_decorator.R
\docType{data}
\name{RutaDecor}
\alias{RutaDecor}
\title{Plot decorators}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RutaDecor
}
\description{
Plot decorators
}
\author{
Zhengjia Wang
}
\keyword{datasets}
