% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_raveplot.R
\docType{data}
\name{RutaPlot}
\alias{RutaPlot}
\title{R6 class for plot class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RutaPlot
}
\description{
R6 class for plot class
}
\author{
zhengjia wang
}
\keyword{datasets}
