% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{scale_01}
\alias{scale_01}
\title{0-1 Scale The Data So We Can Manage The Plot Ranges Easily}
\usage{
scale_01(x)
}
\arguments{
\item{x}{data to be rescaled}
}
\description{
0-1 Scale The Data So We Can Manage The Plot Ranges Easily
}
