% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{trim}
\alias{trim}
\title{Trim Data By Standard Error}
\usage{
trim(x, cutoff = 6)
}
\arguments{
\item{x}{data to be trimmed}

\item{cutoff}{default is 6, then x is clipped +-6 times sd}
}
\description{
Trim Data By Standard Error
}
