% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{trimmed.mse}
\alias{trimmed.mse}
\title{Sd Of Data After Trimmed}
\usage{
trimmed.mse(x, cutoff = 4)
}
\arguments{
\item{x, cutoff}{passed to \link{trim}}
}
\description{
Sd Of Data After Trimmed
}
