% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{add_points}
\alias{add_points}
\title{Same as points, but can be jittered}
\usage{
add_points(x, y, jitr_x = 0, pch = 19, ...)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
}
