% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{clip_x}
\alias{clip_x}
\title{Ensure Data Are Within Some Bounds

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}}
\usage{
clip_x(x, lim)
}
\description{
Ensure Data Are Within Some Bounds

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
