% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{do_if}
\alias{do_if}
\title{Clauses With Side Effects (Plotting Etc)

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}}
\usage{
do_if(boolean_expression, if_clause, else_clause = NULL)
}
\description{
Clauses With Side Effects (Plotting Etc)

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}
}
