% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_data_range}
\alias{get_data_range}
\title{Get data range from a collection of named lists}
\usage{
get_data_range(ll, range_var = "range")
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}
}
