% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{m_se}
\alias{m_se}
\alias{mat_m_se}
\title{Function to return mean and standard error}
\usage{
m_se(x)
mat_m_se(m, DIM=2)

mat_m_se(m, DIM = 2)
}
\description{
Function to return mean and standard error
}
