% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{m_sd}
\alias{m_sd}
\title{Function To Return Mean And Standard Deviation (Na Ignored)

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}}
\usage{
m_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{data}

\item{na.rm}{remove NAs?}
}
\description{
Function To Return Mean And Standard Deviation (Na Ignored)

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
