% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_helpers.R
\name{not_NA}
\alias{not_NA}
\title{Return True If Not Na

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}}
\usage{
not_NA(x)
}
\arguments{
\item{x}{data}
}
\description{
Return True If Not Na

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
