% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_clean}
\alias{plot_clean}
\title{Create a blank plot with given x and y range}
\usage{
plot_clean(xlim, ylim, x = 1, y = 1, type = "n", xlab = "",
  ylab = "", cex.main = rave_cex.main, cex.axis = rave_cex.axis,
  cex.lab = rave_cex.lab, ...)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
\examples{
\dontrun{
# create a blank plot with x from 0 to 10 and y from 1 to 5
plot_clean(0:10, 1:5, xlab = 'X')
}
}
