% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_msg}
\alias{plot_msg}
\title{Show a blank plot with messages}
\usage{
plot_msg(main = "No Conditions Specified")
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}
}
\examples{
\dontrun{
plot_msg("Let's Say Something")
}
}
