% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{rave_barplot}
\alias{rave_barplot}
\title{barplot function that uses all the rave sizes and colors}
\usage{
rave_barplot(height, cex.axis = rave_cex.axis, cex.lab = rave_cex.lab,
  cex.names = rave_cex.lab, ...)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
}
