% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{round_range}
\alias{round_range}
\title{Get a integer interval that contains x}
\usage{
round_range(x)
}
\description{
Get a integer interval that contains x
}
\examples{
\dontrun{
# 0 - 11
round_range(0.5:10.5)
}

\\Sexpr[results=rd, stage=render]{rlang:::lifecycle("maturing")}

}
