% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{stretch}
\alias{stretch}
\title{Useful for plotting when you want to go a bit beyond the data}
\usage{
stretch(x, pct)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}
}
