% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{stretch}
\alias{stretch}
\title{Useful For Plotting When You Want To Go A Bit Beyond The Data

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}}
\usage{
stretch(x, pct)
}
\description{
Useful For Plotting When You Want To Go A Bit Beyond The Data

\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}
}
